Improve naming (2)
