+++ Source/Testably.Expectations/Collections/ThatCollectionExtensions.cs
@@ -1,9 +1,7 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.CompilerServices;
 using Testably.Expectations.Collections;
 using Testably.Expectations.Core;
-using Testably.Expectations.Core.Formatting;
 using Testably.Expectations.Core.Helpers;
 using Testably.Expectations.Core.Results;
 
@@ -13,7 +11,7 @@ namespace Testably.Expectations;
 /// <summary>
 ///     Expectations on collections.
 /// </summary>
-public static class ThatCollectionExtensions
+public static partial class ThatCollectionExtensions
 {
 	/// <summary>
 	///     Verifies that all items in the collection...
@@ -62,29 +60,32 @@ public static class ThatCollectionExtensions
 	/// <summary>
 	///     Verifies that the actual collection contains the <paramref name="expected" /> value.
 	/// </summary>
-	public static AndOrExpectationResult<IEnumerable<TItem>, That<IEnumerable<TItem>>> Contains<TItem>(
-		this That<IEnumerable<TItem>> source,
-		TItem expected,
-		[CallerArgumentExpression("expected")] string doNotPopulateThisValue = "")
-		=> new(source.ExpectationBuilder.Add(new ContainsExpectation<TItem>(expected),
+	public static AndOrExpectationResult<IEnumerable<TItem>, That<IEnumerable<TItem>>>
+		Contains<TItem>(
+			this That<IEnumerable<TItem>> source,
+			TItem expected,
+			[CallerArgumentExpression("expected")] string doNotPopulateThisValue = "")
+		=> new(source.ExpectationBuilder.Add(new ContainsConstraint<TItem>(expected),
 				b => b.AppendMethod(nameof(Contains), doNotPopulateThisValue)),
 			source);
 
 	/// <summary>
 	///     Verifies that the actual collection is empty.
 	/// </summary>
-	public static AndOrExpectationResult<IEnumerable<TItem>, That<IEnumerable<TItem>>> IsEmpty<TItem>(
-		this That<IEnumerable<TItem>> source)
-		=> new(source.ExpectationBuilder.Add(new IsEmptyExpectation<TItem>(),
+	public static AndOrExpectationResult<IEnumerable<TItem>, That<IEnumerable<TItem>>>
+		IsEmpty<TItem>(
+			this That<IEnumerable<TItem>> source)
+		=> new(source.ExpectationBuilder.Add(new IsEmptyConstraint<TItem>(),
 				b => b.AppendMethod(nameof(IsEmpty))),
 			source);
 
 	/// <summary>
 	///     Verifies that the actual collection is not empty.
 	/// </summary>
-	public static AndOrExpectationResult<IEnumerable<TItem>, That<IEnumerable<TItem>>> IsNotEmpty<TItem>(
-		this That<IEnumerable<TItem>> source)
-		=> new(source.ExpectationBuilder.Add(new IsNotEmptyExpectation<TItem>(),
+	public static AndOrExpectationResult<IEnumerable<TItem>, That<IEnumerable<TItem>>>
+		IsNotEmpty<TItem>(
+			this That<IEnumerable<TItem>> source)
+		=> new(source.ExpectationBuilder.Add(new IsNotEmptyConstraint<TItem>(),
 				b => b.AppendMethod(nameof(IsNotEmpty))),
 			source);
 
@@ -95,57 +96,5 @@ public static class ThatCollectionExtensions
 	{
 		source.ExpectationBuilder.AppendExpression(b => b.AppendMethod(nameof(None)));
 		return new QuantifiableCollection<TItem>(source, Quantifier.None);
-	}
-
-	private readonly struct ContainsExpectation<TItem>(TItem expected)
-		: IExpectation<IEnumerable<TItem>>
-	{
-		public ConstraintResult IsMetBy(IEnumerable<TItem> actual)
-		{
-			List<TItem> list = actual.ToList();
-			if (list.Contains(expected))
-			{
-				return new ConstraintResult.Success<IEnumerable<TItem>>(list, ToString());
-			}
-
-			return new ConstraintResult.Failure(ToString(), $"found {Formatter.Format(list)}");
-		}
-
-		public override string ToString()
-			=> $"contains {Formatter.Format(expected)}";
-	}
-
-	private readonly struct IsEmptyExpectation<TItem> : IExpectation<IEnumerable<TItem>>
-	{
-		public ConstraintResult IsMetBy(IEnumerable<TItem> actual)
-		{
-			List<TItem> list = actual.ToList();
-			if (!list.Any())
-			{
-				return new ConstraintResult.Success<IEnumerable<TItem>>(list, ToString());
-			}
-
-			return new ConstraintResult.Failure(ToString(), $"found {Formatter.Format(list)}");
-		}
-
-		public override string ToString()
-			=> "is empty";
-	}
-
-	private readonly struct IsNotEmptyExpectation<TItem> : IExpectation<IEnumerable<TItem>>
-	{
-		public ConstraintResult IsMetBy(IEnumerable<TItem> actual)
-		{
-			List<TItem> list = actual.ToList();
-			if (list.Any())
-			{
-				return new ConstraintResult.Success<IEnumerable<TItem>>(list, ToString());
-			}
-
-			return new ConstraintResult.Failure(ToString(), "it was");
-		}
-
-		public override string ToString()
-			=> "is not empty";
 	}
 }
