+++ Source/Testably.Expectations/Collections/ThatStringCollectionExtensions.ContainsConstraint.cs
@@ -1,0 +1,27 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Testably.Expectations.Core;
+using Testably.Expectations.Core.Formatting;
+
+// ReSharper disable once CheckNamespace
+namespace Testably.Expectations;
+
+public static partial class ThatStringCollectionExtensions
+{
+	private readonly struct ContainsConstraint(string expected) : IExpectation<IEnumerable<string>>
+	{
+		public ConstraintResult IsMetBy(IEnumerable<string> actual)
+		{
+			List<string> list = actual.ToList();
+			if (list.Contains(expected))
+			{
+				return new ConstraintResult.Success<IEnumerable<string>>(list, ToString());
+			}
+
+			return new ConstraintResult.Failure(ToString(), $"found {Formatter.Format(list)}");
+		}
+
+		public override string ToString()
+			=> $"contains {Formatter.Format(expected)}";
+	}
+}
