+++ Source/Testably.Expectations/Collections/ThatStringCollectionExtensions.cs
@@ -1,8 +1,6 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.CompilerServices;
 using Testably.Expectations.Core;
-using Testably.Expectations.Core.Formatting;
 using Testably.Expectations.Core.Helpers;
 using Testably.Expectations.Core.Results;
 
@@ -12,7 +10,7 @@ namespace Testably.Expectations;
 /// <summary>
 ///     Expectations on <see cref="string" /> collections.
 /// </summary>
-public static class ThatStringCollectionExtensions
+public static partial class ThatStringCollectionExtensions
 {
 	/// <summary>
 	///     Verifies that the actual collection contains the <paramref name="expected" /> value.
@@ -21,24 +19,7 @@ public static class ThatStringCollectionExtensions
 		this That<IEnumerable<string>> source,
 		string expected,
 		[CallerArgumentExpression("expected")] string doNotPopulateThisValue = "")
-		=> new(source.ExpectationBuilder.Add(new ContainsExpectation(expected),
+		=> new(source.ExpectationBuilder.Add(new ContainsConstraint(expected),
 				b => b.AppendMethod(nameof(Contains), doNotPopulateThisValue)),
 			source);
-
-	private readonly struct ContainsExpectation(string expected) : IExpectation<IEnumerable<string>>
-	{
-		public ConstraintResult IsMetBy(IEnumerable<string> actual)
-		{
-			List<string> list = actual.ToList();
-			if (list.Contains(expected))
-			{
-				return new ConstraintResult.Success<IEnumerable<string>>(list, ToString());
-			}
-
-			return new ConstraintResult.Failure(ToString(), $"found {Formatter.Format(list)}");
-		}
-
-		public override string ToString()
-			=> $"contains {Formatter.Format(expected)}";
-	}
 }
