+++ Source/Testably.Expectations/Delegates/ThatDelegate.cs
@@ -3,7 +3,8 @@ using Testably.Expectations.Core;
 using Testably.Expectations.Core.Helpers;
 using Testably.Expectations.Core.Results;
 
-namespace Testably.Expectations.Expectations;
+// ReSharper disable once CheckNamespace
+namespace Testably.Expectations;
 
 /// <summary>
 ///     Expectations on delegate values.
@@ -23,7 +24,7 @@ public abstract partial class ThatDelegate
 	public DelegateExpectationResult<TException> Throws<TException>()
 		where TException : Exception
 		=> new(_expectationBuilder.AddCast(
-			new ThrowsExpectation<TException>(),
+			new ThrowsConstraint<TException>(),
 			b => b.Append('.').Append(nameof(Throws)).Append('<').Append(typeof(TException).Name)
 				.Append(">()")));
 
@@ -33,7 +34,7 @@ public abstract partial class ThatDelegate
 	public DelegateExpectationResult<TException> ThrowsExactly<TException>()
 		where TException : Exception
 		=> new(_expectationBuilder.AddCast(
-			new ThrowsExactlyExpectation<TException>(),
+			new ThrowsExactlyConstraint<TException>(),
 			b => b.Append('.').Append(nameof(ThrowsExactly)).Append('<')
 				.Append(typeof(TException).Name).Append(">()")));
 
@@ -42,6 +43,6 @@ public abstract partial class ThatDelegate
 	/// </summary>
 	public DelegateExpectationResult<Exception> ThrowsException()
 		=> new(_expectationBuilder.AddCast(
-			new ThrowsExpectation<Exception>(),
+			new ThrowsConstraint<Exception>(),
 			b => b.AppendMethod(nameof(ThrowsException))));
 }
